Fix tests

+++ src/frontends/pytorch/src/op/min_max.cpp
@@ -42,7 +42,7 @@ OutputVector translate_max(NodeContext& context) {
     auto values = context.mark_node(std::make_shared<v1::ReduceMax>(x, axes_node, keepdims));
     auto k = context.mark_node(std::make_shared<v0::Constant>(element::i32, Shape{}, 1));
     auto topk = std::make_shared<v3::TopK>(x, k, axis_const, v3::TopK::Mode::MAX, v3::TopK::SortType::NONE);
-    auto indicies = context.mark_node(std::make_shared<v0::Convert>(topk->output(1), element::i32));
+    auto indicies = context.mark_node(std::make_shared<v0::Convert>(topk->output(1), element::i64));
     if (!keepdims) {
         indicies = std::make_shared<v0::Squeeze>(indicies, axes_node);
     }
@@ -71,7 +71,7 @@ OutputVector translate_min(NodeContext& context) {
     auto values = context.mark_node(std::make_shared<v1::ReduceMin>(x, axes_node, keepdims));
     auto k = context.mark_node(std::make_shared<v0::Constant>(element::i32, Shape{}, 1));
     auto topk = std::make_shared<v3::TopK>(x, k, axis_const, v3::TopK::Mode::MIN, v3::TopK::SortType::NONE);
-    auto indicies = context.mark_node(std::make_shared<v0::Convert>(topk->output(1), element::i32));
+    auto indicies = context.mark_node(std::make_shared<v0::Convert>(topk->output(1), element::i64));
     if (!keepdims) {
         indicies = std::make_shared<v0::Squeeze>(indicies, axes_node);
     }
